Update all TARDIS-related plugins
