+++ README.md
@@ -3,14 +3,18 @@
 A Dynmap add-on to display [TARDIS](http://dev.bukkit.org/bukkit-plugins/tardis/) locations on a Bukkit server.
 
 ### Requirements
+
 * TARDIS version 2.8-beta-1 build #654 or higher
 * Dynmap (Dynmap-TARDIS was compiled against Dynmap 1.9.2)
 
 ### Installation
 
-1. Install Dynmap as per the instructions found on the [Dynmap](http://dev.bukkit.org/bukkit-plugins/dynmap/) page. Run the server once to generate the default Dynmap files and folders.
+1. Install Dynmap as per the instructions found on the [Dynmap](http://dev.bukkit.org/bukkit-plugins/dynmap/) page. Run
+   the server once to generate the default Dynmap files and folders.
 2. Drop the Dynmap-TARDIS.jar file in the server plugins folder.
-3. Put the [tardis.png](https://github.com/eccentricdevotion/Dynmap-Tardis/blob/master/tardis.png?raw=true) icon file (![image](https://github.com/eccentricdevotion/Dynmap-Tardis/blob/master/tardis.png?raw=true)) into the `plugins/dynmap/web/tiles/_markers_/` folder.
+3. Put the [tardis.png](https://github.com/eccentricdevotion/Dynmap-Tardis/blob/master/tardis.png?raw=true) icon
+   file (![image](https://github.com/eccentricdevotion/Dynmap-Tardis/blob/master/tardis.png?raw=true)) into
+   the `plugins/dynmap/web/tiles/_markers_/` folder.
 4. Run the command `/dmarker addicon id:tardis newlabel:tardis file:plugins/dynmap/web/tiles/_markers_/tardis.png`
 5. Stop and start the server.
 6. Render the map.
