+++ pom.xml
@@ -1,49 +1,148 @@
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>me.eccentric_nz.dynmaptardis</groupId>
-    <artifactId>Dynmap-TARDIS</artifactId>
-    <version>1.3.3</version>
-    <name>Dynmap-TARDIS</name>
-    <repositories>
-        <repository>
-            <id>Spigot</id>
-            <url>https://hub.spigotmc.org/nexus/content/groups/public/</url>
-        </repository>
-    </repositories>
-    <build>
-        <finalName>Dynmap-TARDIS</finalName>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.1</version>
-                <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-    <dependencies>
-        <dependency>
-            <groupId>org.spigotmc</groupId>
-            <artifactId>spigot-api</artifactId>
-            <version>1.16.4-R0.1-SNAPSHOT</version>
-            <type>jar</type>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.dynmap</groupId>
-            <artifactId>dynmap-api</artifactId>
-            <version>3.1-beta-5</version>
-            <scope>system</scope>
-            <systemPath>${basedir}/lib/dynmap-api-3.1-beta5.jar</systemPath>
-        </dependency>
-        <dependency>
-            <groupId>me.eccentric_nz.TARDIS</groupId>
-            <artifactId>TARDIS</artifactId>
-            <version>4.5.4</version>
-        </dependency>
-    </dependencies>
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xmlns="http://maven.apache.org/POM/4.0.0"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>me.eccentric_nz.dynmaptardis</groupId>
+	<artifactId>Dynmap-TARDIS</artifactId>
+	<version>1.3.3-SNAPSHOT</version>
+	<name>Dynmap-TARDIS</name>
+	<description>A Dynmap add-on to display TARDIS locations on Spigot/Paper servers</description>
+	<url>https://github.com/eccentricdevotion/Dynmap-Tardis</url>
+	<properties>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<maven.compiler.source>16</maven.compiler.source>
+		<maven.compiler.target>16</maven.compiler.target>
+		<java.version>16</java.version>
+	</properties>
+	<repositories>
+		<repository>
+			<id>spigot-repo</id>
+			<url>https://hub.spigotmc.org/nexus/content/repositories/snapshots/</url>
+		</repository>
+		<repository>
+			<id>paper-repo</id>
+			<url>https://papermc.io/repo/repository/maven-public</url>
+		</repository>
+		<repository>
+			<id>tardis-repo</id>
+			<url>http://tardisjenkins.duckdns.org:8080/plugin/repository/everything/</url>
+		</repository>
+		<repository>
+			<id>dynmap-repo</id>
+			<url>https://repo.mikeprimm.com/</url>
+		</repository>
+	</repositories>
+	<profiles>
+		<!-- Profiles are used to detect whether this is a local or Jenkins build and adjust the build number accordingly -->
+		<profile>
+			<id>local</id>
+			<activation>
+				<property>
+					<name>!env.BUILD_NUMBER</name>
+				</property>
+			</activation>
+			<properties>
+				<!-- Set the build number to the date and time of the build -->
+				<project.buildNumber>${timestamp}</project.buildNumber>
+			</properties>
+			<build>
+				<plugins>
+					<plugin>
+						<groupId>org.codehaus.mojo</groupId>
+						<artifactId>buildnumber-maven-plugin</artifactId>
+						<version>1.4</version>
+						<configuration>
+							<revisionOnScmFailure>no.scm.config.in.pom</revisionOnScmFailure>
+						</configuration>
+						<executions>
+							<execution>
+								<id>buildnumber</id>
+								<phase>validate</phase>
+								<goals>
+									<goal>create</goal>
+								</goals>
+							</execution>
+						</executions>
+					</plugin>
+				</plugins>
+			</build>
+		</profile>
+		<profile>
+			<id>jenkins</id>
+			<activation>
+				<property>
+					<name>env.BUILD_NUMBER</name>
+				</property>
+			</activation>
+			<properties>
+				<!-- Get the build number from Jenkins -->
+				<project.buildNumber>${env.BUILD_NUMBER}</project.buildNumber>
+			</properties>
+			<build>
+				<plugins>
+					<plugin>
+						<groupId>org.apache.maven.plugins</groupId>
+						<artifactId>maven-javadoc-plugin</artifactId>
+						<version>3.3.0</version>
+						<configuration>
+							<javadocExecutable>${java.home}/../bin/javadoc</javadocExecutable>
+						</configuration>
+						<executions>
+							<execution>
+								<id>javadoc-jar</id>
+								<phase>verify</phase>
+								<goals>
+									<goal>jar</goal>
+								</goals>
+							</execution>
+						</executions>
+					</plugin>
+				</plugins>
+			</build>
+		</profile>
+	</profiles>
+	<build>
+		<finalName>${project.artifactId}-${project.version}-${project.buildNumber}</finalName>
+		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.8.1</version>
+				<configuration>
+					<source>${maven.compiler.source}</source>
+					<target>${maven.compiler.target}</target>
+					<encoding>${project.build.sourceEncoding}</encoding>
+				</configuration>
+			</plugin>
+		</plugins>
+	</build>
+	<dependencies>
+		<dependency>
+			<groupId>org.spigotmc</groupId>
+			<artifactId>spigot-api</artifactId>
+			<version>1.16.5-R0.1-SNAPSHOT</version>
+			<type>jar</type>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>me.eccentric_nz.tardis</groupId>
+			<artifactId>TARDIS</artifactId>
+			<version>4.6.2-SNAPSHOT</version>
+			<type>jar</type>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.dynmap</groupId>
+			<artifactId>dynmap-api</artifactId>
+			<version>2.0</version>
+			<scope>compile</scope>
+		</dependency>
+		<dependency>
+			<groupId>com.google.code.gson</groupId>
+			<artifactId>gson</artifactId>
+			<version>2.8.7</version>
+			<type>jar</type>
+			<scope>compile</scope>
+		</dependency>
+	</dependencies>
 </project>
