+++ src/main/resources/plugin.yml
@@ -5,7 +5,7 @@ name: Dynmap-TARDIS
 load: POSTWORLD
 version: 1.3.3
 api-version: 1.16
-website: http://dev.bukkit.org/bukkit-plugins/tardis
-dev-url: http://dev.bukkit.org/bukkit-plugins/tardis
-url: http://dev.bukkit.org/bukkit-plugins/tardis
-depend: [dynmap,TARDIS]
+website: http://tardisjenkins.duckdns.org:8080/job/Dynmap-TARDIS/
+dev-url: http://github.com/eccentricdevotion/Dynmap-Tardis/
+url: http://eccentricdevotion.github.io/TARDIS/
+depend: [ dynmap,TARDIS ]
